requirements.txt
fastapi
uvicorn
pynacl
requests
python-dotenv
